+++ spinner/spinner.go
@@ -23,6 +23,14 @@ import (
 // ////////////////////////////////////////////////////////////////////////////////// //
 
 const (
+	DURATION_SHORT uint8 = iota
+	DURATION_MINI
+	DURATION_SIMPLE
+)
+
+// ////////////////////////////////////////////////////////////////////////////////// //
+
+const (
 	_ACTION_DONE uint8 = iota
 	_ACTION_ERROR
 	_ACTION_SKIP
@@ -56,6 +64,9 @@ var SkipSymbol = "✔ "
 
 // DisableAnimation is global animation off switch flag
 var DisableAnimation = false
+
+// DurationFormat is format used for printing result action duration
+var DurationFormat = DURATION_SHORT
 
 // ////////////////////////////////////////////////////////////////////////////////// //
 
@@ -150,7 +161,7 @@ func showSpinner() {
 		mu.RLock()
 		fmtc.Printf(spinnerColorTag+"%s  {!}", spinnerFrames[i])
 		fmtc.Print(desc + "… ")
-		fmtc.Printf(timeColorTag+"[%s]{!}", timeutil.ShortDuration(time.Since(start)))
+		fmtc.Printf(timeColorTag+"[%s]{!}", timeutil.Pretty(time.Since(start)).Short(false))
 		mu.RUnlock()
 
 		i++
@@ -195,11 +206,22 @@ func stopSpinner(action uint8) {
 	}
 
 	fmtc.Print(desc + " ")
-	fmtc.Printfn(timeColorTag+"(%s){!}", timeutil.ShortDuration(time.Since(start), true))
+	fmtc.Printfn(timeColorTag+"(%s){!}", formatDuration(time.Since(start)))
 
 	mu.RUnlock()
 
 	mu.Lock()
 	desc, start = "", time.Time{}
 	mu.Unlock()
+}
+
+func formatDuration(d time.Duration) string {
+	switch DurationFormat {
+	case DURATION_MINI:
+		return timeutil.Pretty(d).Mini("")
+	case DURATION_SIMPLE:
+		return timeutil.Pretty(d).Simple()
+	}
+
+	return timeutil.Pretty(d).Short(true)
 }